feat(cactus-web): allow DataGrid layout customization via sub-components
CACTUS-397

BREAKING CHANGE: New DataGrid API structure
