+++ modules/cactus-web/src/DataGrid/DataGrid.story.tsx
@@ -1,8 +1,8 @@
 import { boolean, select, text } from '@storybook/addon-knobs'
 import { storiesOf } from '@storybook/react'
-import React, { ReactElement, useState } from 'react'
+import React, { ReactElement, useContext, useState } from 'react'
 
-import ScreenSizeProvider from '../ScreenSizeProvider/ScreenSizeProvider'
+import { ScreenSizeContext, SIZES } from '../ScreenSizeProvider/ScreenSizeProvider'
 import SplitButton from '../SplitButton/SplitButton'
 import DataGrid from './DataGrid'
 
@@ -81,11 +81,18 @@ const DataGridContainer = ({
   initialData: Record<string, any>[]
   includePaginationAndSort?: boolean
 }): ReactElement => {
+  const cardBreakpoint = select(
+    'cardBreakpoint',
+    ['tiny', 'small', 'medium', 'large', 'extraLarge'],
+    'tiny'
+  )
+  const size = useContext(ScreenSizeContext)
+  const isCardView = cardBreakpoint && size <= SIZES[cardBreakpoint]
   const showResultsCount = includePaginationAndSort
     ? boolean('Show Results Count', true)
     : undefined
-  const providePageSizeOptions = includePaginationAndSort
-    ? boolean('Provide Page Size Options', true)
+  const providePageSizeSelect = includePaginationAndSort
+    ? boolean('Provide Page Size Select', true)
     : undefined
   const providePageCount = includePaginationAndSort
     ? boolean('Provide Page Count', true)
@@ -103,8 +110,7 @@ const DataGridContainer = ({
     currentPage: number
     pageSize: number
     pageCount?: number
-    pageSizeOptions?: number[]
-  }>({ currentPage: 1, pageCount: 3, pageSize: 4, pageSizeOptions: [4, 6, 12] })
+  }>({ currentPage: 1, pageCount: 3, pageSize: 4 })
 
   const clone = (rowData: { [key: string]: any }): void => {
     setData((currentData): { [key: string]: any }[] => [
@@ -184,10 +190,9 @@ const DataGridContainer = ({
     currentPage: number
     pageSize: number
     pageCount?: number
-    pageSizeOptions?: number[]
   } => {
     const paginationOptsCopy = JSON.parse(JSON.stringify(paginationOptions))
-    if (!providePageSizeOptions) {
+    if (!providePageSizeSelect) {
       paginationOptsCopy.pageSizeOptions = undefined
     }
 
@@ -198,47 +203,31 @@ const DataGridContainer = ({
     return paginationOptsCopy
   }
 
+  const usableData = includePaginationAndSort ? paginateData() : data
+  const topSection = boolean('topSection', true)
+
   return (
-    <ScreenSizeProvider>
-      <DataGrid
-        dividers={boolean('dividers', false)}
-        data={includePaginationAndSort ? paginateData() : data}
-        sortOptions={includePaginationAndSort ? sortOptions : undefined}
-        onSort={onSort}
-        paginationOptions={includePaginationAndSort ? getPaginationOptions() : undefined}
-        onPageChange={onPageChange}
-        fullWidth={boolean('fullWidth', false)}
-        cardBreakpoint={select(
-          'cardBreakpoint',
-          ['tiny', 'small', 'medium', 'large', 'extraLarge'],
-          'tiny'
-        )}
-        resultsCountText={showResultsCount ? getResultsCountText() : undefined}
-        pageSizeSelectLabel={includePaginationAndSort ? text('pageSizeSelectLabel', '') : undefined}
-        paginationProps={
-          includePaginationAndSort
-            ? {
-                label: text('Pagination: label', ''),
-                currentPageLabel: text('Pagination: currentPageLabel', ''),
-                prevPageLabel: text('Pagination: prevPageLabel', ''),
-                nextPageLabel: text('Pagination: nextPageLabel', ''),
-                lastPageLabel: text('Pagination: lastPageLabel', ''),
-              }
-            : undefined
-        }
-        prevNextProps={
-          disableNext
-            ? {
-                prevText,
-                nextText,
-                disableNext,
-              }
-            : {
-                prevText,
-                nextText,
-              }
-        }
-      >
+    <DataGrid
+      sortOptions={includePaginationAndSort ? sortOptions : undefined}
+      onSort={onSort}
+      paginationOptions={includePaginationAndSort ? getPaginationOptions() : undefined}
+      onPageChange={onPageChange}
+      fullWidth={boolean('fullWidth', false)}
+      cardBreakpoint={cardBreakpoint}
+    >
+      {topSection && (
+        <DataGrid.TopSection>
+          {showResultsCount && !isCardView && <span>{getResultsCountText()}</span>}
+          {includePaginationAndSort && providePageSizeSelect && (
+            <DataGrid.PageSizeSelect
+              pageSizeOptions={[4, 6, 12]}
+              pageSizeSelectLabel={text('pageSizeSelectLabel', '')}
+              ml={isCardView && size.toString() === 'tiny' ? undefined : 'auto'}
+            />
+          )}
+        </DataGrid.TopSection>
+      )}
+      <DataGrid.Table data={includePaginationAndSort ? paginateData() : data}>
         <DataGrid.DataColumn id="name" title="Name" />
         <DataGrid.DataColumn id="created" title="Created" sortable={sortableCols} />
         <DataGrid.DataColumn
@@ -269,8 +258,37 @@ const DataGridContainer = ({
             </SplitButton>
           )}
         </DataGrid.Column>
-      </DataGrid>
-    </ScreenSizeProvider>
+      </DataGrid.Table>
+      <DataGrid.BottomSection>
+        {isCardView && showResultsCount && size.toString() !== 'tiny' ? (
+          <span>{getResultsCountText()}</span>
+        ) : null}
+        {includePaginationAndSort ? (
+          providePageCount ? (
+            <DataGrid.Pagination
+              label={text('Pagination: label', '')}
+              currentPageLabel={text('Pagination: currentPageLabel', '')}
+              prevPageLabel={text('Pagination: prevPageLabel', '')}
+              nextPageLabel={text('Pagination: nextPageLabel', '')}
+              lastPageLabel={text('Pagination: lastPageLabel', '')}
+              mb={isCardView && size.toString() === 'tiny' ? 4 : undefined}
+              ml={isCardView && size.toString() === 'tiny' ? undefined : 'auto'}
+            />
+          ) : (
+            <DataGrid.PrevNext
+              prevText={prevText}
+              nextText={nextText}
+              disableNext={usableData.length < getPaginationOptions().pageSize || disableNext}
+              mb={isCardView && size.toString() === 'tiny' ? 4 : undefined}
+              ml={isCardView && size.toString() === 'tiny' ? 'auto' : 'auto'}
+            />
+          )
+        ) : null}
+        {isCardView && showResultsCount && size.toString() === 'tiny' ? (
+          <span>{getResultsCountText()}</span>
+        ) : null}
+      </DataGrid.BottomSection>
+    </DataGrid>
   )
 }
